# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AccessLocationSource = typeof AccessLocationSource[keyof typeof AccessLocationSource];


export const AccessLocationSource = {
  ip: 'ip',
  local: 'local',
  unavailable: 'unavailable',
} as const;

export interface AccessLocation {
  source: AccessLocationSource;
  city: string;
  region: string;
  regionCode: string;
  country: string;
  countryCode: string;
  timezone: string;
}

export type PopPersonStatus = typeof PopPersonStatus[keyof typeof PopPersonStatus];


export const PopPersonStatus = {
  titular: 'titular',
  candidato: 'candidato',
} as const;

export interface PopPersonCategory {
  id: string;
  name: string;
  slug: string;
  /** @nullable */
  parentId: string | null;
}

export interface PopPerson {
  name: string;
  category: PopPersonCategory;
  categoryPath: PopPersonCategory[];
  /**
     * Gender recorded for the person, when known.
     * @nullable
     */
  gender: string | null;
  cidade: string;
  estado: string;
  estadoCodigo: string;
  pais: string;
  paisCodigo: string;
  status: PopPersonStatus;
  value: number;
  color: string;
  /**
     * Optional person image URL. The client keeps the cell color as the fallback background.
     * @nullable
     */
  imageUrl: string | null;
}

export type PopPersonElementGender = typeof PopPersonElementGender[keyof typeof PopPersonElementGender];


export const PopPersonElementGender = {
  m: 'm',
  f: 'f',
} as const;

export interface PopPersonElement {
  id: string;
  emoji: string;
  /**
     * Optional item image URL. The client can fall back to emoji when absent.
     * @nullable
     */
  imageUrl: string | null;
  label: string;
  /**
     * Optional item description.
     * @nullable
     */
  description: string | null;
  force: number;
  price: number;
  gender: PopPersonElementGender;
}

export interface PopPersonLevel {
  key: string;
  label: string;
  powerLabel: string;
  emoji: string;
  /** @minimum 0 */
  startDelayMs: number;
  /** @minimum 1 */
  count: number;
  /** @minimum 0 */
  staggerMs: number;
  /** @minimum 0 */
  duration: number;
  growthPerHit: number;
  impactMultiplier: number;
  shake: boolean;
}

export interface PopPersonActionRule {
  elementId: string;
  level: string;
  /** @minimum 0 */
  startDelayMs: number;
  /** @minimum 1 */
  count: number;
  /** @minimum 0 */
  staggerMs: number;
  /** @minimum 0 */
  duration: number;
  growthPerHit: number;
  impactMultiplier: number;
  /** @minimum 0 */
  price: number;
  shake: boolean;
}

export type PopPersonConfigElements = {
  atacar: PopPersonElement[];
  defender: PopPersonElement[];
};

export interface PopPersonConfig {
  elements: PopPersonConfigElements;
  levels: PopPersonLevel[];
  actionRules: PopPersonActionRule[];
}

export type PopPersonActionMode = typeof PopPersonActionMode[keyof typeof PopPersonActionMode];


export const PopPersonActionMode = {
  atacar: 'atacar',
  defender: 'defender',
} as const;

export type PopPersonActionStatus = typeof PopPersonActionStatus[keyof typeof PopPersonActionStatus];


export const PopPersonActionStatus = {
  queued: 'queued',
  running: 'running',
  completed: 'completed',
} as const;

export interface PopPersonAction {
  id: string;
  mode: PopPersonActionMode;
  elementId: string;
  level: string;
  /** @minimum 0 */
  startDelayMs: number;
  targetName: string;
  status: PopPersonActionStatus;
  /** @minimum 0 */
  executeAt: number;
  /** @minimum 0 */
  completesAt: number;
  /**
     * @minimum 0
     * @nullable
     */
  startedAt: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  completedAt: number | null;
  /** @minimum 0 */
  hitCount: number;
  /**
     * @minimum 0
     * @nullable
     */
  lastHitAt: number | null;
  /** @minimum 1 */
  count: number;
  growthPerHit: number;
  impactMultiplier: number;
  /** @minimum 0 */
  staggerMs: number;
  /** @minimum 0 */
  duration: number;
  /** @minimum 0 */
  price: number;
  shake: boolean;
  element: PopPersonElement;
}

export interface PopPersonState {
  /**
     * Monotonic room version used to reject stale realtime snapshots.
     * @minimum 0
     */
  stateVersion: number;
  dataset: PopPerson[];
  actions: PopPersonAction[];
}

export interface AuthenticatedUser {
  xUserId: string;
  username: string;
  name: string;
  /** @nullable */
  xLocation: string | null;
  /** @nullable */
  avatarUrl: string | null;
  /** @nullable */
  email: string | null;
}

export interface PlayerMembership {
  isPlayer: boolean;
  /** @nullable */
  name: string | null;
}

export interface PopPersonBootstrap {
  config: PopPersonConfig;
  state: PopPersonState;
  user: AuthenticatedUser | null;
  player: PlayerMembership;
}

export interface JoinPopPersonResponse {
  player: PopPerson;
}

export interface PlayerRegistrationLocation {
  /** @minLength 1 */
  city: string;
  /** @minLength 1 */
  region: string;
  /** @minLength 1 */
  regionCode: string;
  /** @minLength 1 */
  country: string;
  /** @minLength 1 */
  countryCode: string;
}

export interface JoinPopPersonBody {
  /** @minLength 1 */
  categoryId: string;
  location: PlayerRegistrationLocation;
}

export type PlayerRegistrationLocationValidation = typeof PlayerRegistrationLocationValidation[keyof typeof PlayerRegistrationLocationValidation];


export const PlayerRegistrationLocationValidation = {
  match: 'match',
  different: 'different',
  unknown: 'unknown',
} as const;

export interface PlayerRegistration {
  user: AuthenticatedUser;
  accessLocation: AccessLocation;
  locationValidation: PlayerRegistrationLocationValidation;
  categories: PopPersonCategory[];
  locations: PlayerRegistrationLocation[];
  /** @nullable */
  defaultCategoryId: string | null;
}

export type PopPersonActionInputMode = typeof PopPersonActionInputMode[keyof typeof PopPersonActionInputMode];


export const PopPersonActionInputMode = {
  atacar: 'atacar',
  defender: 'defender',
} as const;

export interface PopPersonActionInput {
  mode: PopPersonActionInputMode;
  elementId: string;
  /** @minLength 1 */
  level: string;
  targetName: string;
  /**
     * @minLength 1
     * @maxLength 160
     */
  idempotencyKey?: string;
}

export interface ErrorResponse {
  error: string;
}

export type StartXAuthenticationParams = {
returnTo?: string;
};

export type GetAuthenticatedUser200 = {
  user: AuthenticatedUser | null;
};