# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pop-person/player:
    post:
      operationId: joinPopPersonAsPlayer
      tags: [pop-person]
      summary: Add the authenticated user as a player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinPopPersonBody"
      responses:
        "201":
          description: Player cell created or restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JoinPopPersonResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pop-person/player/registration:
    get:
      operationId: getPlayerRegistration
      tags: [pop-person]
      summary: Get authenticated player registration defaults
      responses:
        "200":
          description: Registration data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerRegistration"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /pop-person/actions:
    post:
      operationId: createPopPersonAction
      tags: [pop-person]
      summary: Queue an attack or defense action
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PopPersonActionInput"
      responses:
        "201":
          description: Queued action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PopPersonAction"
        "400":
          description: Invalid action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    AccessLocation:
      type: object
      required: [source, city, region, regionCode, country, countryCode, timezone]
      properties:
        source:
          type: string
          enum: [ip, local, unavailable]
        city:
          type: string
        region:
          type: string
        regionCode:
          type: string
        country:
          type: string
        countryCode:
          type: string
        timezone:
          type: string
    PopPerson:
      type: object
      required: [name, category, categoryPath, gender, cidade, estado, estadoCodigo, pais, paisCodigo, status, value, color, imageUrl]
      properties:
        name:
          type: string
        category:
          $ref: "#/components/schemas/PopPersonCategory"
        categoryPath:
          type: array
          items:
            $ref: "#/components/schemas/PopPersonCategory"
        gender:
          type: ["string", "null"]
          description: Gender recorded for the person, when known.
        cidade:
          type: string
        estado:
          type: string
        estadoCodigo:
          type: string
        pais:
          type: string
        paisCodigo:
          type: string
        status:
          type: string
          enum: [titular, candidato]
        value:
          type: number
        color:
          type: string
        imageUrl:
          type: ["string", "null"]
          description: Optional person image URL. The client keeps the cell color as the fallback background.
    PopPersonCategory:
      type: object
      required: [id, name, slug, parentId]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        parentId:
          type: ["string", "null"]
    PopPersonElement:
      type: object
      required: [id, emoji, imageUrl, label, description, force, price, gender]
      properties:
        id:
          type: string
        emoji:
          type: string
        imageUrl:
          type: ["string", "null"]
          description: Optional item image URL. The client can fall back to emoji when absent.
        label:
          type: string
        description:
          type: ["string", "null"]
          description: Optional item description.
        force:
          type: number
        price:
          type: number
        gender:
          type: string
          enum: [m, f]
...
    PopPersonAction:
      type: object
      required: [id, mode, elementId, level, targetName, status, startDelayMs, executeAt, completesAt, startedAt, completedAt, hitCount, lastHitAt, count, growthPerHit, impactMultiplier, staggerMs, duration, price, shake, element]
      properties:
        id:
          type: string
        mode:
          type: string
          enum: [atacar, defender]
        elementId:
          type: string
        level:
          type: string
        startDelayMs:
          type: number
          minimum: 0
        targetName:
          type: string
        status:
          type: string
          enum: [queued, running, completed]
        executeAt:
          type: number
          minimum: 0
        completesAt:
          type: number
          minimum: 0
        startedAt:
          type: ["number", "null"]
          minimum: 0
        completedAt:
          type: ["number", "null"]
          minimum: 0
        hitCount:
          type: number
          minimum: 0
        lastHitAt:
          type: ["number", "null"]
          minimum: 0
        count:
          type: number
          minimum: 1
        growthPerHit:
          type: number
        impactMultiplier:
          type: number
        staggerMs:
          type: number
          minimum: 0
        duration:
          type: number
          minimum: 0
        price:
          type: number
          minimum: 0
        shake:
          type: boolean
        element:
          $ref: "#/components/schemas/PopPersonElement"
...
    JoinPopPersonResponse:
      type: object
      required: [player]
      properties:
        player:
          $ref: "#/components/schemas/PopPerson"
    JoinPopPersonBody:
      type: object
      required: [categoryId, location]
      properties:
        categoryId:
          type: string
          minLength: 1
        location:
          $ref: "#/components/schemas/PlayerRegistrationLocation"
    PlayerRegistrationLocation:
      type: object
      required: [city, region, regionCode, country, countryCode]
      properties:
        city:
          type: string
          minLength: 1
        region:
          type: string
          minLength: 1
        regionCode:
          type: string
          minLength: 1
        country:
          type: string
          minLength: 1
        countryCode:
          type: string
          minLength: 1
    PlayerRegistration:
      type: object
      required: [user, accessLocation, locationValidation, categories, locations, defaultCategoryId]
      properties:
        user:
          $ref: "#/components/schemas/AuthenticatedUser"
        accessLocation:
          $ref: "#/components/schemas/AccessLocation"
        locationValidation:
          type: string
          enum: [match, different, unknown]
        categories:
          type: array
          items:
            $ref: "#/components/schemas/PopPersonCategory"
        locations:
          type: array
          items:
            $ref: "#/components/schemas/PlayerRegistrationLocation"
        defaultCategoryId:
          type: ["string", "null"]
    AuthenticatedUser:
      type: object
      required: [xUserId, username, name, xLocation, avatarUrl, email]
      properties:
        xUserId:
          type: string
        username:
          type: string
        name:
          type: string
        xLocation:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
    PopPersonActionInput:
      type: object
      required: [mode, elementId, level, targetName]
      properties:
        mode:
          type: string
          enum: [atacar, defender]
        elementId:
          type: string
        level:
          type: string
          minLength: 1
        targetName:
          type: string
        idempotencyKey:
          type: string
          minLength: 1
          maxLength: 160
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string