# lib/api-zod/src/generated/types/authenticatedUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthenticatedUser {
  xUserId: string;
  username: string;
  name: string;
  /** @nullable */
  xLocation: string | null;
  /** @nullable */
  avatarUrl: string | null;
  /** @nullable */
  email: string | null;
}