# lib/api-zod/src/generated/types/joinPopPersonBody.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PlayerRegistrationLocation } from './playerRegistrationLocation';

export interface JoinPopPersonBody {
  /** @minLength 1 */
  categoryId: string;
  location: PlayerRegistrationLocation;
}