# lib/api-zod/src/generated/types/playerRegistration.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AccessLocation } from './accessLocation';
import type { AuthenticatedUser } from './authenticatedUser';
import type { PlayerRegistrationLocation } from './playerRegistrationLocation';
import type { PlayerRegistrationLocationValidation } from './playerRegistrationLocationValidation';
import type { PopPersonCategory } from './popPersonCategory';

export interface PlayerRegistration {
  user: AuthenticatedUser;
  accessLocation: AccessLocation;
  locationValidation: PlayerRegistrationLocationValidation;
  categories: PopPersonCategory[];
  locations: PlayerRegistrationLocation[];
  /** @nullable */
  defaultCategoryId: string | null;
}