# lib/api-zod/src/generated/types/playerRegistrationLocation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerRegistrationLocation {
  /** @minLength 1 */
  city: string;
  /** @minLength 1 */
  region: string;
  /** @minLength 1 */
  regionCode: string;
  /** @minLength 1 */
  country: string;
  /** @minLength 1 */
  countryCode: string;
}